fix: change money and outflow fields from int to float in backend schemas and models to support decimal numbers

+++ src/API/app/domain/models/account.py
@@ -11,7 +11,7 @@ class Account:
     id: str
     name: str
     type: str
-    starting_balance: int
+    starting_balance: float
     start_date: str
     yield_rate: float
     assigned_income_source_ids: list[str]
